Software Rendered 3D Update Added Xcode-iOS project with SDL2, for iOS building Added COLLADAReader by Jimita Too many changes to list

+++ include/Engine/Hashing/CRC32.h
@@ -15,8 +15,8 @@ class CRC32 {
 public:
     static Uint32 EncryptString(char* data);
     static Uint32 EncryptString(const char* message);
-    static Uint32 EncryptData(const void* data, Uint32 size);
-    static Uint32 EncryptData(const void* data, Uint32 size, Uint32 crc);
+    static Uint32 EncryptData(const void* data, size_t size);
+    static Uint32 EncryptData(const void* data, size_t size, Uint32 crc);
 };
 
 #endif /* ENGINE_HASHING_CRC32_H */
