+++ include/Engine/Hashing/CombinedHash.h
@@ -15,7 +15,7 @@ class CombinedHash {
 public:
     static Uint32 EncryptString(char* data);
     static Uint32 EncryptString(const char* message);
-    static Uint32 EncryptData(const void* message, Uint32 len);
+    static Uint32 EncryptData(const void* message, size_t len);
 };
 
 #endif /* ENGINE_HASHING_COMBINEDHASH_H */
