+++ include/Engine/Hashing/FNV1A.h
@@ -15,8 +15,8 @@ class FNV1A {
 public:
     static Uint32 EncryptString(char* message);
     static Uint32 EncryptString(const char* message);
-    static Uint32 EncryptData(const void* data, Uint32 size);
-    static Uint32 EncryptData(const void* data, Uint32 size, Uint32 hash);
+    static Uint32 EncryptData(const void* data, size_t size);
+    static Uint32 EncryptData(const void* data, size_t size, Uint32 hash);
 };
 
 #endif /* ENGINE_HASHING_FNV1A_H */
