+++ include/Engine/Hashing/MD5.h
@@ -18,7 +18,7 @@ private:
 public:
     static Uint8* EncryptString(Uint8* dest, char* message);
     static Uint8* EncryptString(Uint8* dest, const char* message);
-    static Uint8* EncryptData(Uint8* dest, void* data, Uint32 size);
+    static Uint8* EncryptData(Uint8* dest, void* data, size_t size);
 };
 
 #endif /* ENGINE_HASHING_MD5_H */
